Moved solve_system to generic ProblemDAE class
