+++ pySDC/projects/DAE/problems/synchronous_machine.py
@@ -1,7 +1,6 @@
 import numpy as np
 import warnings
 from scipy.interpolate import interp1d
-from scipy.optimize import root
 
 from pySDC.projects.DAE.misc.ProblemDAE import ptype_dae
 from pySDC.core.Problem import WorkCounter
@@ -147,8 +146,7 @@ class synchronous_machine_infinite_bus(ptype_dae):
         Defines the mechanical torque applied to the rotor shaft.
     work_counters : WorkCounter
         Counts the work, i.e., number of function calls of right-hand side is called and stored in
-        ``work_counters['rhs']``, the number of function calls of the Newton-like solver is stored in
-        ``work_counters['newton']``.
+        ``work_counters['rhs']``.
 
     References
     ----------
@@ -191,7 +189,6 @@ class synchronous_machine_infinite_bus(ptype_dae):
         self.T_m = 0.854
 
         self.work_counters['rhs'] = WorkCounter()
-        self.work_counters['newton'] = WorkCounter()
 
     def eval_f(self, u, du, t):
         r"""
@@ -268,36 +265,6 @@ class synchronous_machine_infinite_bus(ptype_dae):
         )
         self.work_counters['rhs']()
         return f
-
-    def solve_system(self, impl_sys, u0, t):
-        r"""
-        Solver for nonlinear implicit system (defined in sweeper).
-
-        Parameters
-        ----------
-        impl_sys : callable
-            Implicit system to be solved.
-        u0 : dtype_u
-            Initial guess for solver.
-        t : float
-            Current time :math:`t`.
-
-        Returns
-        -------
-        me : dtype_u
-            Numerical solution.
-        """
-
-        me = self.dtype_u(self.init)
-        opt = root(
-            impl_sys,
-            u0,
-            method='hybr',
-            tol=self.newton_tol,
-        )
-        me[:] = opt.x
-        self.work_counters['newton'].niter += opt.nfev
-        return me
 
     def u_exact(self, t):
         """
